Adjust logging, improve post time handling:
Turns out one reason that advisories can be released but not posted is
because the service I'm using can be >5 minutes behind, so we can't just
rely on run interval to determine if we haven't posted something. So now
we do two things:

* check if post is within interval time, because otherwise stuff won't
  get posted on start-up
* check if a post is within startup time and not marked as a known
  posted ID

Between these two things, we should be able to pull the latest
advisories, have some new ones be on there but "too old" and still post
them.