Add support for message chunking, handling post errors.
App was crashing when post failed, this is my attempt to make it more
robust. Even if the chunking doesn't fix the issue, I should at least
get graceful errors when posting fails.

Calling this 1.1.0 since it supports a new env var and the way that
messages are posted has changed in a notable way.