Handle banner alerts, fix nullable period problem
Now processing banner alerts as service alerts since it looks like they
can be a big deal (track closure) but may not be posted as regular
service alerts. Little bit of a shot in the dark on this one, will have
to wait for wild banner alerts to confirm behavior.

Active periods are nullable, which makes little sense but whatever.
Updated the model and made appropriate changes. Also taking active
period end into account for when start is empty.

End result for both of these changes is that more things will get
posted, and in certain situations can be posted more than once. As long
as that'll only happen once per instance cycle, I think that's okay.